# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: KoalaCast REST API
  description: |
    Privacy-first podcast player API. Unix timestamps are explicitly identified as
    either seconds or milliseconds. Durations and playback positions use milliseconds.
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Server root
security: []
...
paths:
...
  /api/v1/auth/account:
    delete:
      tags: [Authentication]
      operationId: deleteAccount
      summary: Permanently delete the account and every record belonging to it
      description: >-
        Irreversible. Requires the account password or a recovery code in the
        body even though the request is already authenticated, so a borrowed
        unlocked device cannot delete an account. Every dependent row is removed
        in one transaction and all sessions and device credentials are revoked.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: At least one credential is required; clients may populate both fields.
              properties:
                password:
                  type: string
                recovery_code:
                  type: string
      responses:
        '204':
          description: Account deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/export:
    get:
      tags: [Authentication]
      operationId: exportAccount
      summary: Download every record stored for the account
      description: >-
        Returns account metadata, subscriptions, favorites and the latest
        synchronized payload per entity. Never includes password hashes,
        recovery-code hashes, session secrets or device-token hashes.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Account export
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth/data:
    delete:
      tags: [Authentication]
      operationId: deleteSynchronizedData
      summary: Delete synchronized data while retaining the account and logins
      description: >-
        Requires the account password or recovery code again. In one transaction,
        advances the account data generation, disables Global Stats consent and
        deletes subscriptions, favorites, playback, listening, history, queues,
        synchronized settings, sync metadata and Web Push subscriptions. Identity,
        role, sessions and device credentials remain. Stale sync pushes carrying an
        older generation are rejected.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: At least one credential is required; clients may populate both fields.
              properties:
                password:
                  type: string
                recovery_code:
                  type: string
      responses:
        '200':
          description: Synchronized data deleted; account retained
          content:
            application/json:
              schema:
                type: object
                required: [data_generation]
                properties:
                  data_generation:
                    type: integer
                    format: int64
                    minimum: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: koalacast_session
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or credentials invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: string
...